+++ packages/orq-rc/RELEASES.md
@@ -2299,3 +2299,13 @@ Based on:
 - [typescript v3.8.0-rc.36] packages/orq-rc
 ### Releases
 - [NPM v3.8.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.36 - packages/orq-rc
+
+## 2025-06-22 19:01:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.568.0 (2.632.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.9.0-rc.0] packages/orq-rc
+### Releases
+- [NPM v3.9.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/3.9.0-rc.0 - packages/orq-rc